+++ pkg/controller/config.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -13,6 +12,12 @@ import (
 	"github.com/suzuki-shunsuke/go-template-unmarshaler/text"
 	"gopkg.in/yaml.v2"
 )
+
+type Package struct {
+	Name     string `validate:"required"`
+	Registry string `validate:"required"`
+	Version  string `validate:"required"`
+}
 
 type Config struct {
 	Packages       []*Package   `validate:"dive"`
@@ -33,37 +38,14 @@ func (pkgInfos *PackageInfos) UnmarshalYAML(unmarshal func(interface{}) error) e
 	}
 	list := make([]PackageInfo, len(arr))
 	for i, p := range arr {
-		switch p.Type {
-		case pkgInfoTypeGitHubRelease:
-			pkgInfo := &GitHubReleasePackageInfo{
-				Name:        p.Name,
-				RepoOwner:   p.RepoOwner,
-				RepoName:    p.RepoName,
-				Asset:       p.Asset,
-				ArchiveType: p.ArchiveType,
-				Files:       p.Files,
-			}
-			list[i] = pkgInfo
-		case pkgInfoTypeHTTP:
-			pkgInfo := &HTTPPackageInfo{
-				Name:        p.Name,
-				ArchiveType: p.ArchiveType,
-				URL:         p.URL,
-				Files:       p.Files,
-			}
-			list[i] = pkgInfo
-		default:
-			return errors.New("type is invalid")
+		pkgInfo, err := p.GetPackageInfo()
+		if err != nil {
+			return err
 		}
+		list[i] = pkgInfo
 	}
 	*pkgInfos = list
 	return nil
-}
-
-type Package struct {
-	Name     string `validate:"required"`
-	Registry string `validate:"required"`
-	Version  string `validate:"required"`
 }
 
 type PackageInfo interface {
@@ -87,149 +69,27 @@ type mergedPackageInfo struct {
 	URL         *text.Template
 }
 
-type GitHubReleasePackageInfo struct {
-	Name        string  `validate:"required"`
-	ArchiveType string  `yaml:"archive_type"`
-	Files       []*File `validate:"required,dive"`
-
-	RepoOwner string         `yaml:"repo_owner" validate:"required"`
-	RepoName  string         `yaml:"repo_name" validate:"required"`
-	Asset     *text.Template `validate:"required"`
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetName() string {
-	return pkgInfo.Name
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetType() string {
-	return pkgInfoTypeGitHubRelease
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetArchiveType() string {
-	return pkgInfo.ArchiveType
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetPkgPath(rootDir string, pkg *Package) (string, error) {
-	assetName, err := pkgInfo.RenderAsset(pkg)
-	if err != nil {
-		return "", fmt.Errorf("render the asset name: %w", err)
+func (pkgInfo *mergedPackageInfo) GetPackageInfo() (PackageInfo, error) {
+	switch pkgInfo.Type {
+	case pkgInfoTypeGitHubRelease:
+		return &GitHubReleasePackageInfo{
+			Name:        pkgInfo.Name,
+			RepoOwner:   pkgInfo.RepoOwner,
+			RepoName:    pkgInfo.RepoName,
+			Asset:       pkgInfo.Asset,
+			ArchiveType: pkgInfo.ArchiveType,
+			Files:       pkgInfo.Files,
+		}, nil
+	case pkgInfoTypeHTTP:
+		return &HTTPPackageInfo{
+			Name:        pkgInfo.Name,
+			ArchiveType: pkgInfo.ArchiveType,
+			URL:         pkgInfo.URL,
+			Files:       pkgInfo.Files,
+		}, nil
+	default:
+		return nil, errors.New("type is invalid")
 	}
-	return filepath.Join(rootDir, "pkgs", pkgInfo.GetType(), "github.com", pkgInfo.RepoOwner, pkgInfo.RepoName, pkg.Version, assetName), nil
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetFiles() []*File {
-	return pkgInfo.Files
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) GetFileSrc(pkg *Package, file *File) (string, error) {
-	assetName, err := pkgInfo.RenderAsset(pkg)
-	if err != nil {
-		return "", fmt.Errorf("render the asset name: %w", err)
-	}
-	if isUnarchived(pkgInfo.GetArchiveType(), assetName) {
-		return assetName, nil
-	}
-	if file.Src == nil {
-		return file.Name, nil
-	}
-	src, err := file.RenderSrc(pkg, pkgInfo)
-	if err != nil {
-		return "", fmt.Errorf("render the template file.src: %w", err)
-	}
-	return src, nil
-}
-
-func (pkgInfo *GitHubReleasePackageInfo) RenderAsset(pkg *Package) (string, error) {
-	return pkgInfo.Asset.Execute(map[string]interface{}{ //nolint:wrapcheck
-		"Package":     pkg,
-		"PackageInfo": pkgInfo,
-		"OS":          runtime.GOOS,
-		"Arch":        runtime.GOARCH,
-	})
-}
-
-type HTTPPackageInfo struct {
-	Name        string  `validate:"required"`
-	ArchiveType string  `yaml:"archive_type"`
-	Files       []*File `validate:"required,dive"`
-
-	URL *text.Template `validate:"required"`
-}
-
-func (pkgInfo *HTTPPackageInfo) GetName() string {
-	return pkgInfo.Name
-}
-
-func (pkgInfo *HTTPPackageInfo) GetType() string {
-	return pkgInfoTypeHTTP
-}
-
-func (pkgInfo *HTTPPackageInfo) GetArchiveType() string {
-	return pkgInfo.ArchiveType
-}
-
-func (pkgInfo *HTTPPackageInfo) GetPkgPath(rootDir string, pkg *Package) (string, error) {
-	uS, err := pkgInfo.URL.Execute(map[string]interface{}{
-		"Package":     pkg,
-		"PackageInfo": pkgInfo,
-		"OS":          runtime.GOOS,
-		"Arch":        runtime.GOARCH,
-	})
-	if err != nil {
-		return "", fmt.Errorf("render URL: %w", err)
-	}
-	u, err := url.Parse(uS) // TODO
-	if err != nil {
-		return "", fmt.Errorf("parse the URL: %w", err)
-	}
-	return filepath.Join(rootDir, "pkgs", pkgInfo.GetType(), u.Host, u.Path), nil
-}
-
-func (pkgInfo *HTTPPackageInfo) GetFiles() []*File {
-	return pkgInfo.Files
-}
-
-func (pkgInfo *HTTPPackageInfo) GetFileSrc(pkg *Package, file *File) (string, error) {
-	assetName, err := pkgInfo.RenderAsset(pkg)
-	if err != nil {
-		return "", fmt.Errorf("render the asset name: %w", err)
-	}
-	if isUnarchived(pkgInfo.GetArchiveType(), assetName) {
-		return assetName, nil
-	}
-	if file.Src == nil {
-		return file.Name, nil
-	}
-	src, err := file.RenderSrc(pkg, pkgInfo)
-	if err != nil {
-		return "", fmt.Errorf("render the template file.src: %w", err)
-	}
-	return src, nil
-}
-
-func (pkgInfo *HTTPPackageInfo) RenderURL(pkg *Package) (string, error) {
-	uS, err := pkgInfo.URL.Execute(map[string]interface{}{
-		"Package":     pkg,
-		"PackageInfo": pkgInfo,
-		"OS":          runtime.GOOS,
-		"Arch":        runtime.GOARCH,
-	})
-	if err != nil {
-		return "", fmt.Errorf("render URL: %w", err)
-	}
-	return uS, nil
-}
-
-func (pkgInfo *HTTPPackageInfo) RenderAsset(pkg *Package) (string, error) {
-	uS, err := pkgInfo.RenderURL(pkg)
-	if err != nil {
-		return "", fmt.Errorf("render URL: %w", err)
-	}
-	u, err := url.Parse(uS) // TODO
-	if err != nil {
-		return "", fmt.Errorf("parse the URL: %w", err)
-	}
-	return filepath.Base(u.Path), nil
 }
 
 type File struct {
